requirements.txt
email-validator
sendgrid
redis
flask
python-dotenv